Ajuste de condições

+++ scripts/stop-spark.sh
@@ -3,4 +3,7 @@
 $SPARK_HOME/sbin/stop-all.sh
 
 pid=$(pgrep -f jupyter)
-kill -9 $pid
+
+if [ ! -z "$pid" ]; then
+    kill -9 $pid
+fi
